extract components file
