+++ apps/web/src/pages/ApplicantDashboardPage/components/FunctionalCommunityListItem.tsx
@@ -1,13 +1,10 @@
-import { JSX } from "react";
 import { useIntl } from "react-intl";
-import BriefcaseIcon from "@heroicons/react/20/solid/BriefcaseIcon";
-import PresentationChartLineIcon from "@heroicons/react/20/solid/PresentationChartLineIcon";
-import XCircleIcon from "@heroicons/react/20/solid/XCircleIcon";
-import ExclamationTriangleIcon from "@heroicons/react/20/solid/ExclamationTriangleIcon";
 
 import { FragmentType, getFragment, graphql } from "@gc-digital-talent/graphql";
 import { HeadingLevel, PreviewList } from "@gc-digital-talent/ui";
 import { commonMessages } from "@gc-digital-talent/i18n";
+
+import { MetaDataJobInterest, MetaDataTrainingInterest } from "./iconElements";
 
 const PreviewListItemFunctionalCommunity_Fragment = graphql(/* GraphQL */ `
   fragment PreviewListItemFunctionalCommunity on CommunityInterest {
@@ -44,113 +41,6 @@ const FunctionalCommunityListItem = ({
     functionalCommunityListItemQuery,
   );
 
-  const sharedIconStyling = {
-    "data-h2-height": "base(x1)",
-    "data-h2-width": "base(x1)",
-    "data-h2-display": "base(inline-block)",
-    "data-h2-vertical-align": "base(bottom)",
-    "data-h2-margin-right": "base(x.25)",
-    "data-h2-padding-top": "base(x.125)",
-  };
-
-  const interestedWork = (
-    <span>
-      <BriefcaseIcon data-h2-color="base(success)" {...sharedIconStyling} />
-      {intl.formatMessage({
-        defaultMessage: "Interested in work",
-        id: "1VKNrs",
-        description: "Phrase marking interest in community work opportunities",
-      })}
-    </span>
-  );
-  const notInterestedWork = (
-    <span>
-      <XCircleIcon data-h2-color="base(gray.lighter)" {...sharedIconStyling} />
-      {intl.formatMessage({
-        defaultMessage: "Not interested in work",
-        id: "VDVRPt",
-        description:
-          "Phrase marking lack of interest in community work opportunities",
-      })}
-    </span>
-  );
-  const missingWork = (
-    <span>
-      <ExclamationTriangleIcon
-        data-h2-color="base(error)"
-        {...sharedIconStyling}
-      />
-      <span data-h2-color="base(error.darker) base:dark(error.lightest)">
-        {intl.formatMessage({
-          defaultMessage: "Missing work info",
-          id: "mT1G4k",
-          description:
-            "Phrase marking incomplete community work opportunities interest",
-        })}
-      </span>
-    </span>
-  );
-
-  const generateMetaDataJobInterest = (
-    jobInterest: boolean | null | undefined,
-  ): JSX.Element => {
-    if (jobInterest === null || jobInterest === undefined) {
-      return missingWork;
-    }
-    return jobInterest ? interestedWork : notInterestedWork;
-  };
-
-  const interestedTraining = (
-    <span>
-      <PresentationChartLineIcon
-        data-h2-color="base(success)"
-        {...sharedIconStyling}
-      />
-      {intl.formatMessage({
-        defaultMessage: "Interested in training",
-        id: "ERsZAD",
-        description:
-          "Phrase marking interest in community training opportunities",
-      })}
-    </span>
-  );
-  const notInterestedTraining = (
-    <span>
-      <XCircleIcon data-h2-color="base(gray.lighter)" {...sharedIconStyling} />
-      {intl.formatMessage({
-        defaultMessage: "Not interested in training",
-        id: "8wU0cq",
-        description:
-          "Phrase marking lack of interest in community training opportunities",
-      })}
-    </span>
-  );
-  const missingTraining = (
-    <span>
-      <ExclamationTriangleIcon
-        data-h2-color="base(error)"
-        {...sharedIconStyling}
-      />
-      <span data-h2-color="base(error.darker) base:dark(error.lightest)">
-        {intl.formatMessage({
-          defaultMessage: "Missing training info",
-          id: "zGma0A",
-          description:
-            "Phrase marking incomplete community training opportunities interest",
-        })}
-      </span>
-    </span>
-  );
-
-  const generateMetaDataTrainingInterest = (
-    trainingInterest: boolean | null | undefined,
-  ): JSX.Element => {
-    if (trainingInterest === null || trainingInterest === undefined) {
-      return missingTraining;
-    }
-    return trainingInterest ? interestedTraining : notInterestedTraining;
-  };
-
   type MetaDataProps = React.ComponentProps<
     typeof PreviewList.Item
   >["metaData"];
@@ -159,15 +49,21 @@ const FunctionalCommunityListItem = ({
     {
       key: "job-interest",
       type: "text",
-      children: generateMetaDataJobInterest(
-        functionalCommunityListItemFragment.jobInterest,
+      children: (
+        <MetaDataJobInterest
+          jobInterest={functionalCommunityListItemFragment.jobInterest}
+        />
       ),
     },
     {
       key: "training-interest",
       type: "text",
-      children: generateMetaDataTrainingInterest(
-        functionalCommunityListItemFragment.trainingInterest,
+      children: (
+        <MetaDataTrainingInterest
+          trainingInterest={
+            functionalCommunityListItemFragment.trainingInterest
+          }
+        />
       ),
     },
   ];
